requirements.txt
absl-py==1.4.0
asgiref==3.6.0
astunparse==1.6.3
axios==0.3.0
blinker==1.6.2
bottle==0.12.25
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
distlib==0.3.6
Django==4.2.1
filelock==3.12.0
Flask==2.3.2
Flask-Cors==3.0.10
flatbuffers==23.3.3
fonttools==4.39.3
gast==0.4.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.0
h5py==3.8.0
idna==3.4
itsdangerous==2.1.2
jax==0.4.8
Jinja2==3.1.2
joblib==1.2.0
keras==2.12.0
kiwisolver==1.4.4
libclang==16.0.0
lxml==4.9.2
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
ml-dtypes==0.1.0
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
platformdirs==3.5.0
protobuf==4.22.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
requests==2.29.0
requests-oauthlib==1.3.1
rich==13.3.5
rsa==4.9
scikit-learn==1.1.3
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
sklearn-porter==0.7.4
sqlparse==0.4.4
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.1.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
virtualenv==20.23.0
Werkzeug==2.3.3
wrapt==1.14.1
